Move operations into modules reflecting their names (append, remove)
