+++ vczstore/__init__.py
@@ -1,1 +1,3 @@
-from .zarr_impl import append, remove  # noqa F401
+from .append import append  # noqa F401
+from .normalise import normalise  # noqa F401
+from .remove import remove  # noqa F401
